requirements.txt
pytest
torch==1.13.1
torchaudio
torchvision
pytorch_lightning==1.8.6
lightning-flash==0.8.1
torchmetrics==0.9.3
albumentations
wandb
hydra-core
opencv-python
natsort
librosa
scikit-learn==1.3.2
lmdb
matplotlib
streamlit
soundfile